# backend/openapi/case-platform.openapi.yaml
openapi: 3.1.0
info:
  title: LA Industrial Computer Case Platform API
  version: 1.1.0
  description: Multi-case execution, identity, indexed manuals, governed graph and work-order API.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/platform/case-runs:
    post:
      summary: Create an isolated CaseRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/WriteContract"
                - type: object
                  required: [caseId, input]
                  properties:
                    caseId:
                      type: string
                      pattern: "^CASE-[A-Z0-9-]+$"
                    input:
                      type: object
                      additionalProperties: true
      responses:
        "201":
          description: Created run at revision 1
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/platform/auth/login:
    post:
      tags: [Identity]
      summary: Authenticate a platform user and issue an opaque session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account, password]
              properties:
                account: {type: string}
                password: {type: string, format: password, writeOnly: true}
      responses:
        "200":
          description: Session token returned once; only its digest is persisted
        "401":
          description: Invalid credentials
        "423":
          description: Account temporarily locked
...
  /api/platform/admin/users:
    get:
      tags: [Identity]
      summary: Page and filter platform users
      security: [{BearerAuth: []}]
      responses:
        "200": {description: User page without password or token hashes}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      tags: [Identity]
      summary: Create an engineer, expert or administrator
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201": {description: Created public user projection}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /api/platform/admin/audit-events:
    get:
      tags: [Audit]
      summary: Page immutable, redacted business audit events
      security: [{BearerAuth: []}]
      responses:
        "200": {description: Filtered audit event page}
        "403": {$ref: "#/components/responses/Forbidden"}
...
  /api/platform/manuals/import:
    post:
      tags: [ManualKnowledge]
      summary: Persist, extract, chunk and index one PDF manual
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, metadata]
              properties:
                file: {type: string, format: binary}
                metadata: {type: string, description: JSON object}
      responses:
        "201": {description: Indexed manual and import report}
        "409": {description: Content hash already imported}
        "413": {description: File exceeds configured limit}
        "415": {description: File is not a PDF}
        "422": {description: PDF is encrypted, invalid or has no text}
...
  /api/platform/manuals/search:
    post:
      tags: [ManualKnowledge]
      summary: Search page-level manual evidence with FTS5 and normalized fallback
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualSearch"
      responses:
        "200": {description: Ranked evidence with document hash and page reference}
...
  /api/platform/graph/diff:
    get:
      tags: [GovernedGraph]
      summary: Compare added, updated and removed graph entities
      security: [{BearerAuth: []}]
      parameters:
        - {in: query, name: fromVersionId, required: true, schema: {type: string}}
        - {in: query, name: toVersionId, schema: {type: string}}
      responses:
        "200": {description: Node and edge version difference}
  /api/platform/graph/change-sets:
    get:
      tags: [GovernedGraph]
      summary: List governed graph change sets
      security: [{BearerAuth: []}]
      responses:
        "200": {description: Role-filtered change sets}
    post:
      tags: [GovernedGraph]
      summary: Create a draft based on the current graph version
      security: [{BearerAuth: []}]
      responses:
        "201": {description: Draft change set}
...
  /api/platform/work-orders:
    get:
      tags: [WorkOrders]
      summary: Page work orders with role filtering
      security: [{BearerAuth: []}]
      responses:
        "200": {description: Work order page}
    post:
      tags: [WorkOrders]
      summary: Create one work order for an existing CaseRun
      security: [{BearerAuth: []}]
      responses:
        "201": {description: Versioned work order}
        "409": {description: CaseRun already has a work order}
...
  /api/platform/work-orders/{orderId}/job-cards:
    get:
      tags: [WorkOrders]
      summary: List immutable job-card document versions
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/OrderId"}
      responses:
        "200": {description: Job-card metadata versions}
    post:
      tags: [WorkOrders]
      summary: Snapshot current CaseRun data and render a server-side PDF
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/OrderId"}
      responses:
        "201": {description: New or idempotently reused job-card version}
...
  /api/platform/job-cards/{documentId}/download:
    get:
      tags: [WorkOrders]
      summary: Verify SHA-256 and download an archived PDF
      security: [{BearerAuth: []}]
      parameters:
        - {$ref: "#/components/parameters/JobCardDocumentId"}
      responses:
        "200":
          description: Verified PDF document
          content:
            application/pdf:
              schema: {type: string, format: binary}
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Opaque
  parameters:
...
    OrderId:
      in: path
      name: orderId
      required: true
      schema: {type: string}
    JobCardDocumentId:
      in: path
      name: documentId
      required: true
      schema: {type: string}
...
  responses:
...
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Case, run or step does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Revision, state or idempotency conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Actor role is not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    CreateUser:
      type: object
      additionalProperties: false
      required: [account, displayName, role, password]
      properties:
        account:
          type: string
          pattern: "^[a-zA-Z][a-zA-Z0-9_.-]{2,31}$"
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          format: password
          minLength: 10
          maxLength: 128
          writeOnly: true
        profile:
          type: object
          additionalProperties: true
    ManualSearch:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 200
        documentId: {type: string}
        vendor: {type: string}
        equipmentType: {type: string}
        faultDomain: {type: string}
        limit:
          type: integer
          minimum: 1
          maximum: 50
    UserRole:
      type: string
      enum: [engineer, expert, admin]
    Actor:
      type: object
      additionalProperties: false
      required: [id, role]
      properties:
        id:
          type: string
          minLength: 1
        role:
          $ref: "#/components/schemas/UserRole"
    WriteContract:
      type: object
      required: [idempotencyKey, actor]
      properties:
        idempotencyKey:
          type: string
          format: uuid
        actor:
          $ref: "#/components/schemas/Actor"
...
    CaseRun:
      type: object
      required: [runId, caseId, packageVersion, packageHash, status, revision, payload]
      properties:
        runId:
          type: string
        caseId:
          type: string
        packageVersion:
          type: string
        packageHash:
          type: string
        status:
          type: string
          enum: [created, intake_confirmed, diagnosed, plan_confirmed, in_progress, engineer_submitted, expert_reviewing, approved, rejected, published, synced]
        revision:
          type: integer
          minimum: 1
        payload:
          type: object
          additionalProperties: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SuccessEnvelope:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data: {}
    RunResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/CaseRun"
...
    ErrorEnvelope:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true